chore: Move helper function to database package
